+++ test/test_analysis_ran_sev.py
@@ -1,6 +1,6 @@
 
 
-from ec4py.analysis_levich import Levich
+from ec4py.analysis_ran_sev import ran_sev
 from ec4py.util import Quantity_Value_Unit as QVU
 #"import inc_dec    # "The code to test
 import numpy as np
@@ -13,16 +13,16 @@ rot =np.array([300,400,500])
 def  rn(value):
     return round(value*1000000)/1000000
 
-class Test_Analysis_Levich(unittest.TestCase):
+class Test_Analysis_RanSevich(unittest.TestCase):
     
     def test_linear_slope(self):
         
         y_data = 1*np.sqrt(rot)
         unit = "AAA"
-        result = Levich(rot, y_data, unit, "", "bo")
+        result = ran_sev(rot, y_data, unit, "", "bo")
         self.assertEqual(rn(result.value), 1.0)
         y_data = 2*y_data
-        result = Levich(rot, y_data, unit, "", "bo")
+        result = ran_sev(rot, y_data, unit, "", "bo")
         self.assertEqual(rn(result.value), 2.0)
         # self.assertEqual(result.unit, "m")
 
@@ -31,19 +31,19 @@ class Test_Analysis_Levich(unittest.TestCase):
         
         y_data = 1*rot
         unit = "AAA"
-        result = Levich(rot, y_data, "", "", "bo")
+        result = ran_sev(rot, y_data, "", "", "bo")
         self.assertEqual(result.unit, "rpm^-0.5")
-        result = Levich(rot, y_data, unit, "", "bo")
+        result = ran_sev(rot, y_data, unit, "", "bo")
         self.assertEqual(result.unit, unit+" rpm^-0.5")
         # self.assertEqual(result.unit, "m")
     
     def test_quantity_slope(self):
        
         y_data = 1*rot
-        result = Levich(rot, y_data, "", "", "bo")
+        result = ran_sev(rot, y_data, "", "", "bo")
         self.assertEqual(result.quantity, "w^-0.5")
         q = "AAA"
-        result = Levich(rot, y_data, "", q, "bo")
+        result = ran_sev(rot, y_data, "", q, "bo")
         self.assertEqual(result.quantity, q+" w^-0.5")
         
     def real_data(self):